changing double to bigdecimal

+++ apexauto/src/main/java/com/example/apexauto/DTO/VehicleFilterDTO.java
@@ -3,6 +3,8 @@ package com.example.apexauto.DTO;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 // This DTO is used to carry the data for filtering vehicles based on various criteria.
 @Getter
 @Setter
@@ -15,8 +17,8 @@ public class VehicleFilterDTO {
     private Integer year;
     private Integer minYear;
     private Integer maxYear;
-    private Double minPrice;
-    private Double maxPrice;
+    private BigDecimal minPrice;
+    private BigDecimal maxPrice;
     private Boolean isOnSale;
     private Boolean isInStock;
 }
